s/Float61/Float32/

+++ auxiliary.jl
@@ -22,7 +22,7 @@ function trapz(x::Vector, y::Vector)
     return sum * 0.5
 end
 
-mat2gray(input::Image{Float64}) = share(input,
+mat2gray(input::Image{Float32}) = share(input,
     uint8(clamp(data(input) * 255 / maximum(data(input)), 0, 255)))
 
-gray2mat(input::Image{Uint8}) = share(input, float64(input.data / 255))
+gray2mat(input::Image{Uint8}) = share(input, float32(input.data / 255))
