+++ geometry.jl
@@ -5,10 +5,10 @@ immutable Point{T}
     j::T
 end
 
-function interpolate(input::AbstractImage{Float64,2}, p::Point{Float64})
+function interpolate(input::AbstractImage{Float32,2}, p::Point{Float32})
     # Get fractional and integral part of the coordinates
     p_int::Point{Int} = Point(ifloor(p.i), ifloor(p.j))
-    p_fract::Point{Float64} = Point(p.i-p_int.i, p.j-p_int.j)
+    p_fract::Point{Float32} = Point(p.i-p_int.i, p.j-p_int.j)
 
     # Bilinear interpolation
     @inbounds return (input[p_int.i,   p_int.j]   * (1-p_fract.j) * (1-p_fract.i) +
@@ -17,7 +17,7 @@ function interpolate(input::AbstractImage{Float64,2}, p::Point{Float64})
                       input[p_int.i+1, p_int.j+1] * p_fract.j     * p_fract.i)
 end
 
-function resize(input::AbstractImage{Float64,2}, new_size::(Uint, Uint))
+function resize(input::AbstractImage{Float32,2}, new_size::(Uint, Uint))
     # Calculate transform matrix
     transform::Matrix = [
     size(input, 1)/new_size[1]  0;
@@ -48,12 +48,12 @@ function resize(input::AbstractImage{Float64,2}, new_size::(Uint, Uint))
     share(input, output)
 end
 
-function pad(input::AbstractImage{Float64,2})
+function pad(input::AbstractImage{Float32,2})
     origin::Vector = ifloor(([size(input)...] .+ 1) ./ 2)
     rLast::Int = iceil(hypot(([size(input)...] .- 1 - origin)...)) + 1
     rFirst::Int = -rLast
     nBins::Int = rLast - rFirst + 1
-    padded::Array{Float64} = zeros(nBins, nBins)
+    padded::Array{Float32} = zeros(nBins, nBins)
     origin_padded::Vector = ifloor(([size(padded)...] .+ 1) ./ 2)
     offset::Vector = origin_padded - origin
     endpoint::Vector = offset+([size(input)...])
@@ -62,8 +62,8 @@ function pad(input::AbstractImage{Float64,2})
     share(input, padded)
 end
 
-function rotate(input::AbstractImage{Float64,2}, origin::Point{Float64},
-                angle::Float64)
+function rotate(input::AbstractImage{Float32,2}, origin::Point{Float32},
+                angle::Float32)
     if isyfirst(input)
         angle = -angle
     end
